pyproject.toml
[tool.poetry]
name = "useful_fastapi"
version = "0.1.0"
description = ""
authors = ["Ernur Eskermes <eskermesernur86@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.68.1"}
alembic = "^1.7.1"
emails = "^0.6"
PyJWT = "^2.1.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
Authlib = "^0.15.4"
httpx = "^0.19.0"
typer = "^0.4.0"
psycopg2 = "^2.9.1"
psycopg2-binary = "^2.9.1"
tortoise-orm = "^0.17.7"
aerich = "^0.5.8"
asyncpg = "^0.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
